+++ server/src/routes/upload.ts
@@ -1,31 +1,16 @@
 import { Router } from "express";
-import multer from "multer";
+import { upload } from "../multer/multer-config";
+// import multer from "multer";
 
 const router = Router();
 
-const storage = multer.diskStorage({
-  destination: (req, file, cb) => {
-    cb(null, "../bulletinsDeSalaire");
-  },
+//TODO MIDDLEWARE TO CHECK req.body
 
-  filename: (req, file, cb) => {
-    // console.log(file);
-    cb(null, Date.now() + "-" + file.originalname);
-  },
-});
-
-const upload = multer({ storage: storage });
-
-//upload.single("key") "key" doit correspondre a la key du fichier du formData passé
-//Form Data : key : (binary)
-router.post("/upload", upload.single("bulletin"), (req, res) => {
-  // if (!(req.body?.mois && req.body?.user)) {
-  //   console.log("🟥 Mauvaise requête il manque un élément");
-  //   return res.status(400).json("Mauvaise requête");
-  // }
-  // upload.single("bulletin");
-  // console.log({ info: "try", file: req.file, data: req.body });
-  return res.status(200).json({ info: "try", file: req.file, data: req.body });
+router.post("/upload", upload, (req, res) => {
+  // console.log("🚀 ~ file: upload.ts:9 ~ req.body", req.body);
+  return res
+    .status(200)
+    .json({ info: "success", file: req.file, data: req.body });
 });
 
 export { router };
